Revert "adds config to force http to https"
This addition assumes apache which we are not using. The equivalent
change should be made to the nginx stanza if we were using nginx to do
the redirect. However, we're using cloudfront to do the redirect so the
right thing here is probably just to disable port 80 on the elb.

Also, the change accidentally munged the closing brace from the nginx
config over which broke things. oops.

This reverts commit 65cd58d6358b38919684e6e1a760b7125589fbff.