Loosen the contraints to not break existing tests.
Some tests for escaping in the API uses the name field to insert
bogus HTML tags.

Some other tests initialize user with an empty string for a name.

It'd probably be better to just disallow both those cases via
validation, but I'm too tired to update the tests safely.

Loosening the validation still fixes the major escaping flaw and is
strictly better than what's in production.